Create a root reducer in reducer.js
Create a root reducer which will combine only auth reducer for now, but in future other reducers will be combined.
Entities reducer is used to combine reducers of one logical group and for now it will combine only the auth reducer but later it will combine Solution and Issue reducers.